리팩토링
1. 사용하지 않는 코드 삭제
2. 불필요한 캐스팅 삭제

+++ SamFirm/FWFetch.cs
@@ -89,7 +89,7 @@ namespace SamFirm
             int num = 0;
             while ((index-- >= 0) && (num >= 0))
             {
-                num = samMobileHtml.IndexOf("<a class=\"firmware-table-row\" href=\"", (int) (num + 1));
+                num = samMobileHtml.IndexOf("<a class=\"firmware-table-row\" href=\"", num + 1);
             }
             if (num < 0)
             {
